% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_basic_annotations.R
\name{build_basic_annotations}
\alias{build_basic_annotations}
\title{Fetch annotations}
\usage{
build_basic_annotations()
}
\value{
A GRangesList object with a GRanges for every basic annotation.
}
\description{
\code{get_basic_annotations} fetch annotations from UCSC hg19 and from Fantom
}
\details{
This function is meant to be run only to generate the GRangesList
object that will contain the general annotation. It should only be
used before building the package. The goal of this function is to
fetch the basic annotations from UCSC (genes, exons, TSS, TTR) and
enhancers from Fantom.
}
\examples{
\dontrun{
build_basic_annotations()
}
}

