% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scores.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{Calculate the p-value associated with a given overlap percentage.}
\usage{
calculate_pvalue(obs, overlaps, sample_size, sample_count)
}
\arguments{
\item{obs}{The observed percentage of overlap.}

\item{overlaps}{The pre-calculated overlaps between baseline and current
subject. Must be a logical vector of the same length as
baseline.}

\item{sample_size}{The number of value to pick for each draw.}

\item{sample_count}{The number of draw to calculate p-value.}
}
\value{
The p-value for current subject.
}
\description{
Calculate the p-value associated with a given overlap percentage.
}

