% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scores.R
\name{percent_overlap}
\alias{percent_overlap}
\title{Calculate the percentage between 2 GRanges objects.}
\usage{
percent_overlap(qry, sbj)
}
\arguments{
\item{qry}{Query. The percentage will be calculated relatively to this
object. Must of class \code{GRanges}.}

\item{sbj}{Subject. Must be of class \code{GRanges} or \code{GIntervalTree}.}
}
\value{
The percentage of the query that overlaps the subject.
}
\description{
Calculate the percentage between 2 GRanges objects.
}
\examples{
gr1 <- GRanges("chr1", IRanges(1,100))
gr2 <- GRanges("chr1", IRanges(40,140))
percent_overlap(gr1, gr2)
gri2 <- GIntervalTree(gr2)
percent_overlap(gr1, gr2)
}

